requirements.txt
# ============================================================
# DataFlow 2026 — User Behavior Prediction
# Python >= 3.10  |  CUDA >= 11.8 (optional, for GPU training)
# ============================================================
#
# GPU PyTorch — install separately BEFORE this file:
#   pip install torch==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121
#
# CPU-only PyTorch (slow for DL models):
#   pip install torch>=2.0.0
#
# Everything else:
#   pip install -r requirements.txt
# ============================================================


# ── Core Data Science ─────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0              # scipy.stats in statistical_features.py


# ── Machine Learning ──────────────────────────────────────────
scikit-learn>=1.3.0        # TfidfVectorizer, LabelEncoder, metrics, KFold, PCA, KMeans
xgboost>=2.0.0             # XGBoostMultiOutput; GPU via device="cuda"
lightgbm>=4.0.0            # LGBMClassifier with early stopping


# ── Deep Learning ─────────────────────────────────────────────
# torch is installed separately (see CUDA instructions above).
# Uncomment below only for CPU-only environments:
# torch>=2.0.0


# ── Utilities ─────────────────────────────────────────────────
joblib>=1.3.0              # Parallel target training in XGBoostMultiOutput
pyyaml>=6.0.0              # load_config (import yaml)
tqdm>=4.65.0               # Progress bars in training loops


# ── Visualization ─────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.13.0            # Statistical plots in EDA notebook
plotly>=5.15.0             # Interactive plots (notebooks)


# ── Notebook Environment ──────────────────────────────────────
jupyter>=1.0.0
jupyterlab>=4.0.0          # Modern Jupyter UI (replaces classic notebook)
ipykernel>=6.25.0          # Python kernel for Jupyter
ipywidgets>=8.0.0          # Interactive widgets (tqdm notebook integration)
nbformat>=5.9.0            # Notebook file format
nbconvert>=7.0.0           # Convert .ipynb to script / HTML


# ── Testing ───────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0          # Coverage reporting


# ── Code Quality ──────────────────────────────────────────────
black>=23.0.0
isort>=5.12.0


# ── Hyperparameter Tuning (optional) ──────────────────────────
# Uncomment when running Optuna hyperparameter search:
# optuna>=3.4.0
